requirements.txt
pandas
torch
numpy<2.0.0
python-dotenv
matplotlib